Modulized the shortcut mode
