+++ Code/renderer/shortcutMode/constants.js
@@ -1,0 +1,19 @@
+export const FILE_EXTENSIONS = [
+  '.js', '.jsx', '.ts', '.tsx', '.mjs', '.cjs',
+  '.py', '.pyw', '.pyi',
+  '.php', '.phtml',
+  '.html', '.htm', '.css', '.scss', '.sass', '.less',
+  '.c', '.cpp', '.cc', '.cxx', '.h', '.hpp', '.hxx',
+  '.cs',
+  '.java', '.jar', '.class',
+  '.go',
+  '.rs',
+  '.rb',
+  '.swift',
+  '.kt', '.kts',
+  '.dart',
+  '.sh', '.bash', '.zsh', '.bat', '.cmd', '.ps1',
+  '.json', '.xml', '.yaml', '.yml', '.toml', '.ini', '.cfg', '.conf',
+  '.txt', '.md', '.markdown', '.rst',
+  '.sql', '.r', '.m', '.lua', '.pl', '.pm', '.tcl'
+];
